# Conflicts:
#	source/TestSceneHawk.cpp

+++ source/Framebuffer.cpp
@@ -28,7 +28,7 @@ void Framebuffer::addColorTexture(int index) {
 	glGenTextures(1, &(colorTex[index]) );
 	glBindTexture(GL_TEXTURE_2D, colorTex[index]);
 
-	GLint format = (hdrEnabled) ? GL_RGBA16F : GL_RGBA;
+	GLint format = (hdrEnabled) ? GL_RGBA32F : GL_RGBA;
 	glTexImage2D(GL_TEXTURE_2D, 0, format, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, 0);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
